auto_generate_video_fold6.old/services/voice-enhancement/requirements.txt
# Voice Enhancement Service Dependencies

# FastAPI 核心套件
fastapi==0.109.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# 音訊處理核心
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.11.0

# 深度學習框架
torch>=2.2.0
torchaudio>=2.2.0
transformers>=4.35.0

# 語音合成模型
TTS>=0.21.0
espnet>=202301
espnet-model-zoo>=0.1.7

# 音訊增強
noisereduce>=3.0.0
pyworld>=0.3.0
praat-parselmouth>=0.4.3

# 情感識別
speechemotionrecognition>=1.0.0
opensmile>=2.4.2

# 多語言支援
polyglot>=16.7.4
langdetect>=1.0.9

# 音訊格式轉換
pydub>=0.25.1
ffmpeg-python>=0.2.0

# 機器學習工具
scikit-learn>=1.4.0
joblib>=1.3.0

# 語音克隆
resemblyzer>=0.1.1
encoder>=0.1.0

# 資料庫連接
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# HTTP 客戶端
httpx>=0.25.0
aiohttp>=3.9.0

# 身份驗證
python-jose[cryptography]>=3.3.4

# 日誌記錄
structlog>=23.2.0

# 效能監控
prometheus-client>=0.19.0

# 開發工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
requests>=2.32.4
Pillow>=10.3.0
aiohttp>=3.10.11